+++ decidim-accountability/app/presenters/decidim/accountability/result_presenter.rb
@@ -16,14 +16,11 @@ module Decidim
 
       # Render the result title
       #
-      # links - should render hashtags as links?
-      # extras - should include extra hashtags?
-      #
       # Returns a String.
-      def title(links: false, extras: true, html_escape: false, all_locales: false)
+      def title(html_escape: false, all_locales: false)
         return unless result
 
-        super(result.title, links, html_escape, all_locales, extras:)
+        super(result.title, html_escape, all_locales)
       end
     end
   end