+++ decidim-admin/app/controllers/decidim/admin/components_controller.rb
@@ -47,7 +47,7 @@ module Decidim
 
           on(:invalid) do
             flash.now[:alert] = I18n.t("components.create.error", scope: "decidim.admin")
-            render action: "new"
+            render action: "new", status: :unprocessable_entity
           end
         end
       end
@@ -74,7 +74,7 @@ module Decidim
 
           on(:invalid) do
             flash[:alert] = I18n.t("components.update.error", scope: "decidim.admin")
-            render action: :edit
+            render action: :edit, status: :unprocessable_entity
           end
         end
       end