+++ decidim-core/app/packs/src/decidim/editor/index.js
@@ -59,9 +59,6 @@ export default function createEditor(container) {
     };
   }
 
-  if (container.classList.contains("js-hashtags")) {
-    decidimOptions.hashtag = true;
-  }
   if (container.classList.contains("js-mentions")) {
     decidimOptions.mention = true;
   }