+++ decidim-initiatives/app/services/decidim/initiatives/status_change_notifier.rb
@@ -21,7 +21,7 @@ module Decidim
       # * published, discarded: Initiative authors will be notified about the
       #   result of the technical validation process.
       #
-      # * rejected, accepted: Initiative's followers and authors will be
+      # * rejected, accepted: Initiative authors and committee members will be
       #   notified about the result of the initiative.
       def notify
         notify_initiative_creation if initiative.created?
@@ -58,12 +58,6 @@ module Decidim
       end
 
       def notify_support_result
-        initiative.followers.each do |follower|
-          Decidim::Initiatives::InitiativesMailer
-            .notify_state_change(initiative, follower)
-            .deliver_later
-        end
-
         initiative.committee_members.approved.each do |committee_member|
           Decidim::Initiatives::InitiativesMailer
             .notify_state_change(initiative, committee_member.user)