+++ decidim-participatory_processes/app/cells/decidim/participatory_process_groups/content_blocks/main_data_cell.rb
@@ -5,7 +5,6 @@ module Decidim
     module ContentBlocks
       class MainDataCell < Decidim::ContentBlocks::ParticipatorySpaceMainDataCell
         include Decidim::SanitizeHelper
-        include Decidim::TwitterSearchHelper
 
         delegate :title, :description, :group_url, :meta_scope, to: :resource